+++ packages/api/src/app/index.ts
@@ -3,4 +3,4 @@ export * from './config';
 export * from './permissions';
 export * from './cdn';
 export * from './checks';
-export * from './resolveAppConfig';
+export * from './resolve';
